Refactor files to be in the app folder instead of in separate containers folder
